pyproject.toml
[project]
name = "tools-selector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bm25s>=0.2.14",
    "datasets>=4.4.1",
    "docx2txt>=0.9",
    "langchain>=1.0.8",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.0.3",
    "llama-index>=0.14.8",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-openai>=0.6.9",
    "llama-index-llms-openai-like>=0.5.3",
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.2.3",
    "pydantic>=2.12.4",
    "pystemmer>=3.0.0",
    "rank-bm25>=0.2.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
